+++ CHANGELOG.md
@@ -4,12 +4,18 @@
 
 - [FEATURE] custom legend formatting in TimeSeriesChart using `series_name_format` query option #709
 - [FEATURE] User can collapse / expand queries in the panel editor #718
+- [FEATURE] TimeSeriesQuery plugins now accept a `dependsOn` function to optimize loading based on dashboard context #732
 - [ENHANCEMENT] Dashboard variables list is displayed as a sticky header #703
 - [ENHANCEMENT] Disable initial panel animation #710
-- [BUGFIX] Fix Variable data model in the backend to be aligned with what is available in the backend. #734
-- [BREAKINGCHANGE] Variable display configuration has been aligned with panel display configuration. `display.label`
-  becomes `display.name` #734
+- [ENHANCEMENT] time range components styling improvements #733
 - [ENHANCEMENT] Organize time series options in tabs #741
+- [BUGFIX] Fix Variable data model in the backend to be aligned with what is available in the backend. #734
+- [BUGFIX] Back button to dashboard listing broken #735
+- [BREAKINGCHANGE] TimeRangeProvider props changed, `paramsEnabled` added #735
+- [BREAKINGCHANGE] Variable display configuration has been aligned with panel display configuration. `display.label`
+- [BREAKINGCHANGE] Variable plugins `dependsOn` now returns an object #742
+  becomes `display.name` #734
+- [ENHANCEMENT] User always sees edit icons in edit mode (instead of only seeing on hover) #748
 
 ## 0.14.0 / 2022-11-02